# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce Api",
        "version": "0.1"
    },
    "paths": {
...
        "/api/user/address": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "user address store ",
                "description": "Returns user Address data",
                "operationId": "CreateUserAddress",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserAddress"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/user/address/{id}": {
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "user address update ",
                "description": "Returns user Address data",
                "operationId": "updateUserAddress",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "address id field ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserAddress"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/user/addresses": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "user Addresses ",
                "description": "Returns user address data",
                "operationId": "UserAddresses",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/user/cart": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "user cart  ",
                "description": "Returns user cart products",
                "operationId": "UserCart",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "add product to user cart ",
                "description": "adds product to user cart with quantity ",
                "operationId": "addProductToCart",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "inputs product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "merchant_id",
                        "in": "query",
                        "description": "merchant id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "product quantity to add",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/user/cart/remove_product/{id}": {
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "remove product from user cart ",
                "description": "removes product from user cart",
                "operationId": "removeProductFromCart",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "inputs product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/user/favorite": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "add product to user favorite ",
                "description": "adds product to user favorite ",
                "operationId": "addProductToFavorite",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "inputs product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "merchant_id",
                        "in": "query",
                        "description": "merchant id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/user/favorite/{id}": {
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "remove product from user favorite ",
                "description": "removes product from user favorite",
                "operationId": "removeProductFromFavorite",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "inputs product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateUserAddress": {
                "title": " user address request",
                "description": "Some simple request Create user address",
                "properties": {
                    "address": {
                        "title": "address",
                        "description": "address text field",
                        "type": "object",
                        "format": "string",
                        "example": "address details"
                    },
                    "street": {
                        "title": "street",
                        "description": "street text field",
                        "type": "object",
                        "format": "string",
                        "example": "street_details"
                    },
                    "nearest_landmark": {
                        "title": "nearest_landmark",
                        "description": "nearest_landmark field",
                        "type": "string",
                        "format": "string",
                        "example": "nearest_landmark details "
                    },
                    "notes": {
                        "title": "notes",
                        "description": "notes field",
                        "type": "string",
                        "format": "string",
                        "example": "notes details "
                    },
                    "longitude": {
                        "title": "longitude",
                        "description": "longitude ",
                        "type": "string",
                        "format": "float",
                        "example": 30.445
                    },
                    "latitude": {
                        "title": "latitude",
                        "description": "latitude ",
                        "type": "string",
                        "format": "float",
                        "example": 30.145
                    },
                    "mobile": {
                        "title": "mobile",
                        "description": "mobile",
                        "type": "string",
                        "format": "integer",
                        "example": 1066558412
                    },
                    "is_default": {
                        "title": "is_default",
                        "description": "is_default",
                        "type": "string",
                        "format": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "UpdateUserAddress": {
                "title": " user address request",
                "description": "Some simple request update user address",
                "properties": {
                    "address": {
                        "title": "address",
                        "description": "address text field",
                        "type": "object",
                        "format": "string",
                        "example": "address details"
                    },
                    "street": {
                        "title": "street",
                        "description": "street text field",
                        "type": "object",
                        "format": "string",
                        "example": "street_details"
                    },
                    "nearest_landmark": {
                        "title": "nearest_landmark",
                        "description": "nearest_landmark field",
                        "type": "string",
                        "format": "string",
                        "example": "nearest_landmark details "
                    },
                    "notes": {
                        "title": "notes",
                        "description": "notes field",
                        "type": "string",
                        "format": "string",
                        "example": "notes details "
                    },
                    "longitude": {
                        "title": "longitude",
                        "description": "longitude ",
                        "type": "string",
                        "format": "float",
                        "example": 30.445
                    },
                    "latitude": {
                        "title": "latitude",
                        "description": "latitude ",
                        "type": "string",
                        "format": "float",
                        "example": 30.145
                    },
                    "mobile": {
                        "title": "mobile",
                        "description": "mobile",
                        "type": "string",
                        "format": "integer",
                        "example": 1066558412
                    },
                    "is_default": {
                        "title": "is_default",
                        "description": "is_default",
                        "type": "string",
                        "format": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}